@@ -29,63 +29,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.16",
-    "@angular/cdk": "^11.2.6",
-    "@angular/common": "^12.2.16",
-    "@angular/compiler": "^12.2.16",
-    "@angular/core": "^12.2.16",
-    "@angular/fire": "^6.1.4",
-    "@angular/forms": "^12.2.16",
-    "@angular/material": "^11.0.0",
-    "@angular/platform-browser": "^12.2.16",
-    "@angular/platform-browser-dynamic": "^12.2.16",
-    "@angular/router": "^12.2.16",
-    "@angular/service-worker": "^12.2.16",
-    "@fortawesome/angular-fontawesome": "^0.8.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-regular-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@schematics/angular": "^11.2.6",
-    "angular-bootstrap-md": "^11.0.0",
-    "body-parser": "~1.19.0",
-    "core-js": "^3.6.3",
+    "@angular/animations": "^14.0.2",
+    "@angular/cdk": "^14.0.2",
+    "@angular/common": "^14.0.2",
+    "@angular/compiler": "^14.0.2",
+    "@angular/core": "^14.0.2",
+    "@angular/fire": "^7.4.1",
+    "@angular/forms": "^14.0.2",
+    "@angular/material": "^14.0.2",
+    "@angular/platform-browser": "^14.0.2",
+    "@angular/platform-browser-dynamic": "^14.0.2",
+    "@angular/router": "^14.0.2",
+    "@angular/service-worker": "^14.0.2",
+    "@fortawesome/angular-fontawesome": "^0.11.1",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-brands-svg-icons": "^6.1.1",
+    "@fortawesome/free-regular-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@schematics/angular": "^14.0.2",
+    "angular-bootstrap-md": "^13.0.0",
+    "body-parser": "~1.20.0",
+    "core-js": "^3.23.1",
     "crawler-user-agents": "git+https://github.com/monperrus/crawler-user-agents.git",
-    "express": "~4.17.1",
-    "firebase": "^8.3.1",
+    "express": "~4.18.1",
+    "firebase": "^9.8.3",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "method-override": "~3.0.0",
-    "ngx-sharebuttons": "^8.0.5",
-    "rxjs": "^6.5.2",
-    "tslib": "^2.0.0",
+    "ngx-sharebuttons": "^11.0.0",
+    "rxjs": "^7.5.5",
+    "tslib": "^2.4.0",
     "zone.js": "~0.11.6"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.1202.17",
-    "@angular-devkit/build-angular": "~12.2.17",
-    "@angular/cli": "^12.2.17",
-    "@angular/compiler-cli": "^12.2.16",
-    "codelyzer": "^6.0.0",
+    "@angular-devkit/architect": "0.1400.2",
+    "@angular-devkit/build-angular": "~14.0.2",
+    "@angular/cli": "^14.0.2",
+    "@angular/compiler-cli": "^14.0.2",
+    "codelyzer": "^6.0.2",
     "fuzzy": "^0.1.3",
-    "html-webpack-plugin": "^5.3.1",
-    "inquirer": "^6.2.2",
-    "inquirer-autocomplete-prompt": "^1.0.1",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
+    "html-webpack-plugin": "^5.5.0",
+    "inquirer": "^8.2.4",
+    "inquirer-autocomplete-prompt": "^2.0.0",
+    "jasmine-core": "~4.2.0",
+    "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "protractor": "~7.0.0",
-    "ts-node": "^8.5.4",
+    "ts-node": "^10.8.1",
     "tslint": "~6.1.0",
-    "typescript": "^4.3.5",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-middleware": "^4.1.0",
-    "webpack-dev-server": "^3.11.2"
+    "typescript": "^4.7.3",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-middleware": "^5.3.3",
+    "webpack-dev-server": "^4.9.2"
   }
 }
