@@ -1,41 +1,40 @@
-import { Injectable } from '@angular/core';
-import { AngularFireAuth } from '@angular/fire/auth';
-import firebase from 'firebase/app';
-import { AngularFirestore, AngularFirestoreDocument } from '@angular/fire/firestore';
-import { Observable, of } from 'rxjs';
-import { User } from '@wh-objects/auth';
-import { switchMap } from 'rxjs/operators';
+import { Injectable } from "@angular/core";
+import { AngularFireAuth } from "@angular/fire/compat/auth";
+import { GoogleAuthProvider } from "firebase/auth";
+import {
+  AngularFirestore,
+  AngularFirestoreDocument,
+} from "@angular/fire/compat/firestore";
+import { Observable, of } from "rxjs";
+import { User } from "@wh-objects/auth";
+import { switchMap } from "rxjs/operators";
 
 @Injectable()
 export class AuthService {
-
   user: Observable<User>;
 
   constructor(private afAuth: AngularFireAuth, private afs: AngularFirestore) {
-
     //// Get auth data, then get firestore user document || null
-    this.user = this.afAuth.authState
-      .pipe(
-        switchMap(user => {
-          if (user) {
-            return this.afs.doc<User>(`users/${user.uid}`).valueChanges();
-          } else {
-            return of(null);
-          }
-        })
-      );
+    this.user = this.afAuth.authState.pipe(
+      switchMap((user) => {
+        if (user) {
+          return this.afs.doc<User>(`users/${user.uid}`).valueChanges();
+        } else {
+          return of(null);
+        }
+      })
+    );
   }
 
   googleLogin() {
-    const provider = new firebase.auth.GoogleAuthProvider();
+    const provider = new GoogleAuthProvider();
     return this.oAuthLogin(provider);
   }
 
   private oAuthLogin(provider) {
-    return this.afAuth.signInWithPopup(provider)
-      .then((credential) => {
-        this.updateUserData(credential.user);
-      });
+    return this.afAuth.signInWithPopup(provider).then((credential) => {
+      this.updateUserData(credential.user);
+    });
   }
 
   signOut() {
@@ -44,7 +43,9 @@ export class AuthService {
 
   private updateUserData(user) {
     // Sets user data to firestore on login
-    const userRef: AngularFirestoreDocument<any> = this.afs.doc(`users/${user.uid}`);
+    const userRef: AngularFirestoreDocument<any> = this.afs.doc(
+      `users/${user.uid}`
+    );
     const data: User = {
       uid: user.uid,
       displayName: user.displayName,
@@ -52,8 +53,8 @@ export class AuthService {
       roles: {
         admin: false,
         editor: false,
-        sponsor: false
-      }
+        sponsor: false,
+      },
     };
     return userRef.set(data, { merge: true });
   }
@@ -61,25 +62,25 @@ export class AuthService {
   ///// Role-based Authorization //////
 
   canRead(user: User): boolean {
-    const allowed = ['admin', 'editor', 'sponsoring'];
+    const allowed = ["admin", "editor", "sponsoring"];
     return this.checkAuthorization(user, allowed);
   }
 
   canEdit(user: User): boolean {
-    const allowed = ['admin', 'editor', 'sponsoring'];
+    const allowed = ["admin", "editor", "sponsoring"];
     return this.checkAuthorization(user, allowed);
   }
 
   canDelete(user: User): boolean {
-    const allowed = ['admin'];
+    const allowed = ["admin"];
     return this.checkAuthorization(user, allowed);
   }
 
-
-
   // determines if user has matching role
   private checkAuthorization(user: User, allowedRoles: string[]): boolean {
-    if (!user) { return false; }
+    if (!user) {
+      return false;
+    }
     for (const role of allowedRoles) {
       if (user.roles[role]) {
         return true;
@@ -87,5 +88,4 @@ export class AuthService {
     }
     return false;
   }
-
 }
