@@ -1,27 +1,19 @@
-import { Component, OnInit, Input, ElementRef, Renderer2 } from '@angular/core';
-import { NavigationNode } from '../navigation/navigation.service';
-import { AuthService } from '@wh-share/auth.service';
+import { Component, OnInit, Input, ElementRef, Renderer2 } from "@angular/core";
+import { AuthService } from "@wh-share/auth.service";
 
 @Component({
-  selector: 'app-navbar',
-  templateUrl: './navbar.component.html',
-  styleUrls: ['./navbar.component.scss']
+  selector: "app-navbar",
+  templateUrl: "./navbar.component.html",
+  styleUrls: ["./navbar.component.scss"],
 })
-
-
 export class NavbarComponent implements OnInit {
-
-  @Input() nodes: NavigationNode[];
-
   constructor(public auth: AuthService) {
-    console.log('User', auth.user);
+    console.log("User", auth.user);
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
-  onMenuClick() {
-  }
+  onMenuClick() {}
 
   logout() {
     this.auth.signOut();
